feat (plugin): Add inspectPackage function
This is a function to be used along the vite.config.ts file that will read the packages (except dev packages) listed in the package.json and return them to add into shared options of the plugin.
Also the function interate over specific scoped package to find their dependencies as well. This will help the plugin to avoid ship extra code by setting other packages as shared also.